requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
pandas==2.1.0
numpy==1.24.3
python-dotenv==0.19.0
fuzzywuzzy==0.18.0
rapidfuzz==3.6.1
openpyxl==3.0.9
scikit-learn==1.3.0
openai==1.60.0 